pyproject.toml
[project]
name = "ai-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Asanshay Gupta", email = "46036121+SuperAce100@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.93.2",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
